+++ Kernel/CMakeLists.txt
@@ -73,7 +73,7 @@ set(KERNEL_SOURCES
     Graphics/Intel/NativeGraphicsAdapter.cpp
     Graphics/VirtIOGPU/FrameBufferDevice.cpp
     Graphics/VirtIOGPU/Console.cpp
-    Graphics/VirtIOGPU/VirtIOGPU.cpp
+    Graphics/VirtIOGPU/GPU.cpp
     Graphics/VirtIOGPU/GraphicsAdapter.cpp
     Graphics/VGACompatibleAdapter.cpp
     Storage/Partition/DiskPartition.cpp
