+++ Kernel/Graphics/VirtIOGPU/GraphicsAdapter.h
@@ -8,7 +8,7 @@
 
 #include <Kernel/Graphics/VirtIOGPU/Console.h>
 #include <Kernel/Graphics/VirtIOGPU/FrameBufferDevice.h>
-#include <Kernel/Graphics/VirtIOGPU/VirtIOGPU.h>
+#include <Kernel/Graphics/VirtIOGPU/GPU.h>
 
 namespace Kernel::Graphics::VirtIOGPU {
